+++ changelog.md
@@ -4,6 +4,7 @@
 
 - CHANGE: Combine CMD Box (Run Output) and LSP Output into a MessageBox
           bottom of the Editor.
+- CHANGE: Changed "go to line" shortcut from ALT+L to F5.
 
 ## v0.1.0